requirements.txt
anyio==4.1.0
certifi==2023.11.17
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
python-telegram-bot==13.15
sniffio==1.3.0
socksio==1.0.0
emoji==1.7.0